update_daemon/requirements.txt
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
entsoe-py==0.5.9
idna==3.4
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
loguru==0.6.0
paho-mqtt==1.6.1
schedule @ git+https://github.com/dbader/schedule.git@3eac646a8d2658929587d7454bd2c85696df254e
## The following requirements were added by pip freeze:
greenlet==2.0.2
SQLAlchemy==2.0.7
typing_extensions==4.5.0
## The following requirements were added by pip freeze:
mysql-connector-python==8.0.32
protobuf==3.20.3
